+++ gamja/src/pages/mypage/MypageMentee.js
@@ -7,7 +7,7 @@ import MyConcern from "../../components/mypage/MyConcern";
 import JournalList from "../../components/mypage/JournalList";
 import MenteeHistory from "../../components/mypage/MenteeHistory";
 import Column from "../../components/mypage/Column.jsx";
-import LogoutBtn from "../../components/mypage/LogoutBtn.jsx";
+import LogoutWithdrawBtn from "../../components/mypage/LogoutWithdrawBtn.jsx";
 import axios from "axios";
 
 const MypageMentee = () => {
@@ -17,27 +17,52 @@ const MypageMentee = () => {
   const refreshToken = localStorage.getItem("refresh");
 
   function logout() {
-    axios
-      .post(
-        "http://127.0.0.1:8000/users/logout/",
-        {
-          refresh: refreshToken,
-        },
-        {
-          headers: {
-            Authorization: `Bearer ${accessToken}`,
+    if (!window.confirm("로그아웃 하시겠습니까?")) {
+      return;
+    } else {
+      axios
+        .post(
+          "http://127.0.0.1:8000/users/logout/",
+          {
+            refresh: refreshToken,
           },
-        }
-      )
-      .then((response) => {
-        console.log(response.data);
-        localStorage.clear();
-        alert("로그아웃이 완료되었습니다.");
-        navigate("/login");
-      })
-      .catch((error) => {
-        console.log(error);
-      });
+          {
+            headers: {
+              Authorization: `Bearer ${accessToken}`,
+            },
+          }
+        )
+        .then((response) => {
+          console.log(response.data);
+          localStorage.clear();
+          alert("로그아웃이 완료되었습니다.");
+          navigate("/login");
+        })
+        .catch((error) => {
+          console.log(error);
+        });
+    }
+  }
+
+  function menberWithdraw() {
+    if (!window.confirm("탈퇴하시겠습니까?")) {
+      return;
+    } else {
+      axios
+        .delete(
+          `http://127.0.0.1:8000//users/users/${localStorage.getItem(
+            "user_id"
+          )}/`,
+          {
+            headers: { Authorization: `Bearer ${accessToken}` },
+          }
+        )
+        .then((response) => {
+          console.log(response.data);
+          navigate("/login");
+        })
+        .catch();
+    }
   }
 
   useEffect(() => {
@@ -89,7 +114,8 @@ const MypageMentee = () => {
           />
         </ColumnBox>
         <ButtonBox>
-          <LogoutBtn onClick={logout} />
+          <LogoutWithdrawBtn text="logout" onClick={logout} />
+          <LogoutWithdrawBtn text="withdraw" onClick={menberWithdraw} />
         </ButtonBox>
       </Container>
     </>
@@ -106,7 +132,6 @@ const Container = styled.div`
   &::-webkit-scrollbar {
     display: none;
   }
-  padding-bottom: 40px;
 `;
 const Both = styled.div`
   display: flex;
@@ -137,6 +162,13 @@ const ColumnBox = styled.div`
 `;
 
 const ButtonBox = styled.div`
-  margin-top: 31px;
-  margin-left: 438px;
+  display: flex;
+  flex-direction: column;
+  gap: 8px;
+  margin: 32px 40px 20px 0;
+
+  img {
+    margin-left: auto;
+    margin-bottom: auto;
+  }
 `;
